:recycle: Extract ListStylingOptions into own component
